+++ massiv/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fix and export `guardNumberOfElements`
 * `Eq` instances for `IndexException` and `SizeException`
 * Fix `upsample` implementation and improve its performance.
+* Addition of `deleteRegionM`, `deleteRowsM` and `deleteColumnsM`
 
 # 0.3.4